% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/COF.outliers.R
\name{COF.outliers}
\alias{COF.outliers}
\title{Funkcja COF.outliers -  znajduje wartosci odstajace dla danego k}
\usage{
COF.outliers(data_set, k)
}
\arguments{
\item{data_set}{Zbior danych}

\item{k}{k-sasiadow}
}
\description{
Funkcja COF.outliers() dla okreslonego zbioru danych i podanego
wspolczynnika k-sasiadow znajduje wartosci odstajace (outliers)
}
\author{
Czeslaw Horyn choryn@us.edu.pl
}
