% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Collection.R
\name{Collection}
\alias{Collection}
\title{Funkcja Collection - funkcja pomocnicza sluzaca do obliczen}
\usage{
Collection(data_inf2, c)
}
\arguments{
\item{data_inf2}{Zbior danych}

\item{c}{Numer grupy}
}
\description{
Funkcja pomocnicza wykorzystywana w petli zapisuje poszczegolne zbiory
}
\author{
Czeslaw Horyn choryn@us.edu.pl
}
