% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delete.outliers2.R
\name{delete.outliers2}
\alias{delete.outliers2}
\title{Funkcja delete.outliers2 - funkcja usuwajaca odchylenia przed grupowaniem}
\usage{
delete.outliers2(data_inf, o)
}
\arguments{
\item{data_inf}{Zbior danych}

\item{o}{Ilosc odchylen}
}
\description{
Przygotowujemy zbior data_infTrain do dalszego procesu obrobki w kazdej petli usuwajac odpowiednia liczbe odchylen
}
\author{
Czeslaw Horyn choryn@us.edu.pl
}
