% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun.dist.R
\name{fun.dist}
\alias{fun.dist}
\title{Funkcja fun.dist - funkcja obliczajaca odleglosci miedzy obiektami}
\usage{
fun.dist(data_infTrain, methodDist)
}
\arguments{
\item{data_infTrain}{Zbior danych}

\item{methodDist}{Metoda odleglosci}
}
\description{
Funkcja fun.dist() w kazdej petli po usunieciu odpowiednich wierszy z obserwacjami odstajacymi (powstaje plik data_infTrain).
Obliczamy wybrana metoda w funkcji Funk.test odleglosc miedzy obserwacjami podajac metode (parametr methodDist),
otrzymujemy (d_data_infTrain)
}
\author{
Czeslaw Horyn choryn@us.edu.pl
}
