% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Funk.test.R
\name{Funk.test}
\alias{Funk.test}
\title{Wyszukiwanie odchylen i grupowanie danych - Search Outliers and Clustering Raport}
\usage{
Funk.test(file, nameColumns, qSet, methodDist, methodHclust, k.neighbors,
  o.outliers, o.algorithm)
}
\arguments{
\item{file}{Zbior danych}

\item{nameColumns}{Nazwy kolumn}

\item{qSet}{Liczba grup}

\item{methodDist}{Metoda odleglosci}

\item{methodHclust}{Metoda laczenia}

\item{k.neighbors}{Wspolczynnik k-sasiadow}

\item{o.outliers}{Ilosc usuwanych odchylen}

\item{o.algorithm}{Metoda wykrywania odchylen (LOF lub COF)}
}
\value{
Podsumowanie i dendrogram
}
\description{
Funkcja Funk.test() wyszukuje odchylenia metoda LOF lub COF
i nastepnie usuwa je z badanego zbioru, po usunieciu grupuje dane 
algorytmem hierarchicznym. Za kazdym razem obliczny jest wspolczynnik 
Jaacarda, ktory okresla jakosc zbiorow po grupowaniu.
}
\examples{
Funk.test <- function("iris.data",c('ClusterNumber','Sepal length','Sepal width','Petal length','Petal width'),c(1:3),"canberra","ward.D",4,1,"COF")

}
\author{
Czeslaw Horyn choryn@us.edu.pl
}
