% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linkage_cutree.R
\name{linkage_cutree}
\alias{linkage_cutree}
\title{Funkcja linkage_cutree - funkcja laczaca dwa najblizsze skupienia i przycinajaca drzewo (dendrogram)}
\usage{
linkage_cutree(d_data_infTrain, methodHclust, qSet)
}
\arguments{
\item{d_data_infTrain}{Zbior danych}

\item{methodHclust}{Metoda laczenia}

\item{qSet}{Liczba grup}
}
\description{
Funkcja linkage_cutree() w kazdej petli po obliczeniu odleglosci przekazujemy (d_data_infTrain)
i liczymy wybrana metoda (parametr methodHclust) w Funk.test(),
tak otrzymane (hc_data_infTrain) przycinamy wg dlugosci parametru qSet otrzymujac (hc_data_infTrainCutree)
}
\author{
Czeslaw Horyn choryn@us.edu.pl
}
