% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LOF.outliers.R
\name{LOF.outliers}
\alias{LOF.outliers}
\title{Funkcja LOF.outliers -  znajduje wartosci odstajace ustawiajac opcjonalnie k}
\usage{
LOF.outliers(data_set, k)
}
\arguments{
\item{data_set}{Zbior danych}

\item{k}{k-sasiadow}
}
\description{
Funkcja LOF.outliers() dla okreslonego zbioru danych i podanego
wspolczynnika k-sasiadow znajduje wartosci odstajace (outliers)
Funkcja LOF - parametr od k = 1 (nie moze byc 0)
}
\author{
Czeslaw Horyn choryn@us.edu.pl
}
