% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numberInCluster_sumaALL.R
\name{numberInCluster_sumaALL}
\alias{numberInCluster_sumaALL}
\title{Funkcja zliczajaca ilosc obserwacji we wszystkich grupach lacznie}
\usage{
numberInCluster_sumaALL(hc_data_infTrainCutree_)
}
\arguments{
\item{hc_data_infTrainCutree_}{Zbior danych}
}
\description{
Funkcja numberInCluster_sumaALL(hc_data_infTrainCutree_) - oblicza ilosc obserwacji lacznie
}
\author{
Czeslaw Horyn choryn@us.edu.pl
}
