% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numberInCluster_sumaCluster1.R
\name{numberInCluster_sumaCluster1}
\alias{numberInCluster_sumaCluster1}
\title{Funkcja zliczajaca ilosc obserwacji w grupie 1}
\usage{
numberInCluster_sumaCluster1(hc_data_infTrainCutree_)
}
\arguments{
\item{hc_data_infTrainCutree_}{Zbior danych}
}
\description{
Funkcja numberInCluster_sumaALL(hc_data_infTrainCutree_) - oblicza ilosc obserwacji w grupie 1
}
\author{
Czeslaw Horyn choryn@us.edu.pl
}
