% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numberInCluster_sumaCluster2.R
\name{numberInCluster_sumaCluster2}
\alias{numberInCluster_sumaCluster2}
\title{Funkcja zliczajaca ilosc obserwacji w grupie 2}
\usage{
numberInCluster_sumaCluster2(hc_data_infTrainCutree_)
}
\arguments{
\item{hc_data_infTrainCutree_}{Zbior danych}
}
\description{
Funkcja numberInCluster_sumaALL(hc_data_infTrainCutree_) - oblicza ilosc obserwacji w grupie 2
}
\author{
Czeslaw Horyn choryn@us.edu.pl
}
