% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.data.R
\name{read.data}
\alias{read.data}
\title{Funkcja read.data - wczytanie danych (A)}
\usage{
read.data(file_data, nameColumns)
}
\arguments{
\item{file_data}{Zbior danych}

\item{nameColumns}{Nazwy kolumn}
}
\description{
Funkcja read.data() - file -> przygotowane pliki (zbiory danych) nalezy umiescic w folderze 
przygotowanego pakietu R "SOaCRaport (ang. Search Outliers and Clustering Raport)"
Przygotowanie pliku: plik (file,zbior danych) powinien zawierac ponumerowana pierwsza kolumne ClusterNumber i dalej pozostale atrybuty
}
\author{
Czeslaw Horyn choryn@us.edu.pl
}
