% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.data_noClusterNumber.R
\name{read.data_noClusterNumber}
\alias{read.data_noClusterNumber}
\title{Funkcja read.data_noClusterNumber - wczytanie danych bez numerow grup (B)}
\usage{
read.data_noClusterNumber(data_inf_)
}
\arguments{
\item{data_inf_}{Zbior danych}
}
\description{
Funkcja read.data_noClusterNumber() - wczytanie danych bez numerow grup (B)
}
\author{
Czeslaw Horyn choryn@us.edu.pl
}
