% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeInExcel.R
\name{writeInExcel}
\alias{writeInExcel}
\title{Funkcja writeInExcel() zapisujaca dane do pliku Excel o nazwie ZRZUT.xlsx w Arkuszu1}
\usage{
writeInExcel(df_total_)
}
\arguments{
\item{df_total_}{Dane do zapisu}
}
\description{
Funkcja writeInExcel() zapisujaca dane do pliku Excel o nazwie ZRZUT.xlsx w Arkuszu1
}
\author{
Czeslaw Horyn choryn@us.edu.pl
}
